Add support for Read-Composite
Updates CoAP methods to include FETCH, PATCH, and iPATCH. FETCH is
needed for Read-Composite. iPATCH will be needed for Write-Composite.

Minor SenML-JSON parsing and serializing tweaks.

Logic is put in to distinguish Observe-Composite from Read-Composite
but Observe-Composite is not implemented yet.

Read-Composite of / (all objects) may be valid, but is unlikely to
succeed, so is rejected before processing.

A separate command (reaadc) was added to the server rather than just
extending the read command to allow Read-Composite of just a single
URI, which is useful for testing.

Signed-off-by: Scott Bertin <sbertin@telular.com>