Fix memory leak caused by the lwm2m_send function
I recently imported a part of this library, in particular the lwm2m_send(...) function, which is not implemented in the original Eclipse/Wakaama library.

The SEND command works fine, but I noticed a memory leak every time I call this function, I investigated a bit, and I noticed that this is due to the buffer being allocated by the lwm2m_data_serialize(...) (on line 897) function called by the lwm2m_send() is never deallocated.

I see you wrote this comment:
// lwm2m_handle_packet will free buffer
but the lwm2m_handle_packet() function does not deallocate this buffer, so I guess this buffer has to be deallocated manually by us.

I've seen this exact comment elsewhere in the library as well, so I'm guessing you copied and pasted it from the original library, in particular I am referring to the dm_handleRequest(...) and bootstrap_handleCommand() functions respectively on lines 276 and 949. In these two cases it is true that the lwm2m_handle_packet(...) function will free the buffer, precisely on line 342, because this buffer was allocated by the lwm2m_handle_packet() function itself when it called the handle_request() function on line 299, but this is not the case with the lwm2m_send() function.